backend/requirements.txt
numpy
black
flake8
isort
pre-commit

mypy
sphinx
flask
django
notebook
h5py
mysql-connector-python
psycopg2